Update and rename nxn_imitate_if_better
Update - Inserito il datacollector come richiesto
Modificato nome del file
